move metadata decorator and loader to schema package
