+++ master-slave-core/master-slave-core-route/src/main/java/org/apache/shardingsphere/masterslave/metadata/MasterSlaveTableMetaDataLoader.java
@@ -24,7 +24,7 @@ import org.apache.shardingsphere.sql.parser.binder.metadata.table.TableMetaDataL
 import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationProperties;
 import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationPropertyKey;
 import org.apache.shardingsphere.underlying.common.database.type.DatabaseType;
-import org.apache.shardingsphere.underlying.common.metadata.loader.RuleMetaDataLoader;
+import org.apache.shardingsphere.underlying.common.metadata.schema.loader.RuleMetaDataLoader;
 
 import javax.sql.DataSource;
 import java.sql.SQLException;
