+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/core/metadata/ShardingTableMetaDataLoader.java
@@ -29,7 +29,7 @@ import org.apache.shardingsphere.underlying.common.config.properties.Configurati
 import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationPropertyKey;
 import org.apache.shardingsphere.underlying.common.database.type.DatabaseType;
 import org.apache.shardingsphere.underlying.common.exception.ShardingSphereException;
-import org.apache.shardingsphere.underlying.common.metadata.loader.RuleMetaDataLoader;
+import org.apache.shardingsphere.underlying.common.metadata.schema.loader.RuleMetaDataLoader;
 import org.apache.shardingsphere.underlying.common.rule.DataNode;
 
 import javax.sql.DataSource;
